+++ gulp/tasks/default.js
@@ -1,4 +1,0 @@
-var gulp = require('gulp');
-
-// The default task (called when we run `gulp` from cli)
-gulp.task('default', ['watch', 'js']);